+++ .gitignore
@@ -1,6 +1,3 @@
-*node_modules*
-*dist*
-*.cache*
-*.cache*
-*dist*
-*secrets*
+node_modules/
+front/dist/
+front/.cache/
