+++ app/io.js
@@ -1,0 +1,3 @@
+const io = require('socket.io')();
+
+module.exports = io;
