+++ bin/www
@@ -1,0 +1,128 @@
+#!/usr/bin/env node
+
+'use strict';
+
+/**
+ * Module dependencies.
+ */
+
+const app = require('../app');
+const debug = require('debug')('cryptochain:server');
+const http = require('http');
+const superagent = require('superagent');
+const io = require('../app/io');
+
+/**
+ * Get port from environment and store in Express.
+ */
+
+const DEFAULT_PORT = 3000,
+ROOT_NODE_ADDRESS = `http://localhost:${DEFAULT_PORT}`;
+
+let PEER_PORT;
+
+if (process.env.GENERATE_PEER_PORT === 'true') {
+  PEER_PORT = DEFAULT_PORT + Math.ceil(Math.random() * 1000);
+}
+
+const port = normalizePort(process.env.PORT || PEER_PORT || DEFAULT_PORT);
+app.set('port', port);
+app.set('env', process.env.NODE_ENV);
+
+/**
+ * Synchronisation at start with root node
+ */
+
+const syncWithRootState = () => {
+  superagent.get(`${ROOT_NODE_ADDRESS}/api/blocks`).set('accept', 'json').end((err, res) => {
+      if( !err && res.status === 200 ) {
+        console.log('replace chain on a sync with', res.body);
+        app.locals.blockchain.replaceChain(res.body);
+      }
+  });
+
+  superagent.get(`${ROOT_NODE_ADDRESS}/api/transaction-pool-map`).set('accept', 'json').end((err, res) => {
+    if( !err && res.status === 200 ) {
+      console.log('replace transaction pool map on a sync with', res.body);
+      app.locals.transactionPool.setMap(res.body);
+    }
+  });
+}
+/**
+ * Create HTTP server and attach socket.io server to it.
+ */
+
+const server = http.createServer(app);
+io.attach(server);
+
+/**
+ * Listen on provided port, on all network interfaces.
+ */
+
+server.listen(port, () => {
+  if( port !== DEFAULT_PORT ) {
+    syncWithRootState();
+  }
+});
+server.on('error', onError);
+server.on('listening', onListening);
+
+/**
+ * Normalize a port into a number, string, or false.
+ */
+
+function normalizePort(val) {
+  const port = parseInt(val, 10);
+
+  if (isNaN(port)) {
+    // named pipe
+    return val;
+  }
+
+  if (port >= 0) {
+    // port number
+    return port;
+  }
+
+  return false;
+}
+
+/**
+ * Event listener for HTTP server "error" event.
+ */
+
+function onError(error) {
+  if (error.syscall !== 'listen') {
+    throw error;
+  }
+
+  const bind = typeof port === 'string'
+    ? 'Pipe ' + port
+    : 'Port ' + port;
+
+  // handle specific listen errors with friendly messages
+  switch (error.code) {
+    case 'EACCES':
+      console.error(bind + ' requires elevated privileges');
+      process.exit(1);
+      break;
+    case 'EADDRINUSE':
+      console.error(bind + ' is already in use');
+      process.exit(1);
+      break;
+    default:
+      throw error;
+  }
+}
+
+/**
+ * Event listener for HTTP server "listening" event.
+ */
+
+function onListening() {
+  const addr = server.address();
+  const bind = typeof addr === 'string'
+    ? 'pipe ' + addr
+    : 'port ' + addr.port;
+  debug('Listening on ' + bind);
+}
