--- client/src/history.js
@@ -1,3 +1,0 @@
-import createBrowserHistory from 'history/createBrowserHistory';
-
-export default createBrowserHistory();
