+++ config.js
@@ -1,4 +1,7 @@
+'use strict';
+
 const MINE_RATE = 1000;
+
 const INITIAL_DIFFICULTY = 3;
 
 const GENESIS_DATA = {
