+++ front/src/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Cryptochain</title>
+    <link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.min.css">
+  </head>
+  <body>
+    <div id="root"></div>
+    <script src="index.js" type="text/javascript"></script>
+  </body>
+</html>
