+++ front/src/index.js
@@ -1,0 +1,22 @@
+
+import React from 'react';
+import { render } from 'react-dom';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import App from './components/App';
+import Blocks from './components/Blocks';
+import ConductTransaction from './components/ConductTransaction';
+import TransactionPool from './components/TransactionPool';
+import history from './history';
+import './stylesheets/style.css';
+
+render(
+  <Router history={history}>
+    <Switch>
+      <Route exact path='/' component={App}/>
+      <Route path='/blocks' component={Blocks}/>
+      <Route path='/conduct-transaction' component={ConductTransaction}/>
+      <Route path='/transaction-pool' component={TransactionPool}/>
+    </Switch>
+  </Router>,
+  document.getElementById('root')
+);
