--- util/crypto-hash.js
@@ -1,11 +1,0 @@
-const crypto = require('crypto');
-
-const cryptoHash = (...inputs) => {
-  const hash = crypto.createHash('sha256');
-
-  hash.update(inputs.map(input => JSON.stringify(input)).sort().join(' '));
-
-  return hash.digest('hex');
-};
-
-module.exports = cryptoHash;
