+++ wallet/transaction-pool.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const Transaction = require('./transaction');
 
 class TransactionPool {
@@ -5,36 +7,32 @@ class TransactionPool {
     this.transactionMap = {};
   }
 
-  clear() {
-    this.transactionMap = {};
-  }
-
   setTransaction(transaction) {
     this.transactionMap[transaction.id] = transaction;
+  }
+
+  existingTransaction({ inputAddress }) {
+    return Object.values(this.transactionMap).find(transaction => transaction.input.address === inputAddress);
   }
 
   setMap(transactionMap) {
     this.transactionMap = transactionMap;
   }
 
-  existingTransaction({ inputAddress }) {
-    const transactions = Object.values(this.transactionMap);
-
-    return transactions.find(transaction => transaction.input.address === inputAddress);
-  }
-
   validTransactions() {
     return Object.values(this.transactionMap).filter(
-      transaction => Transaction.validTransaction(transaction)
+      transaction => Transaction.validTransaction({ transaction })
     );
   }
 
-  clearBlockchainTransactions({ chain }) {
-    for (let i=1; i<chain.length; i++) {
-      const block = chain[i];
+  clear() {
+    this.transactionMap = {};
+  }
 
-      for (let transaction of block.data) {
-        if (this.transactionMap[transaction.id]) {
+  clearBlockchainTransactions({ chain }) {
+    for ( let i = 0 ; i < chain.length ; i++ ) {
+      for( let transaction of chain[i].data ) {
+        if(this.transactionMap[transaction.id]) {
           delete this.transactionMap[transaction.id];
         }
       }
